pyproject.toml
[tool.poetry]
name = "netsight"
version = "0.1.0"
description = ""
authors = ["wangxin.jeffry <wangxin.jeffry@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.86.0"
SQLAlchemy = "^1.4.43"
PyJWT = "^2.6.0"
orjson = "^3.8.1"
celery = "^5.2.7"
redis = "^4.3.4"
prometheus-fastapi-instrumentator = "^5.9.1"
httpx = "^0.23.0"
gunicorn = "^20.1.0"
uvicorn = "^0.19.0"
loguru = "^0.6.0"
asgi-correlation-id = "^3.2.0"
toml = "^0.10.2"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"
email-validator = "^1.3.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
autoflake = "^1.7.7"
ipykernel = "^6.17.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
